restructure folder and moving workspace on vscode
