+++ lib/app/router.dart
@@ -3,10 +3,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
 // Import halaman-halaman utama
-import '../features/auth/login_page.dart';
-import '../features/admin/admin_dashboard.dart';
-// import '../features/doctor/doctor_dashboard.dart';
-// import '../features/patient/patient_dashboard.dart';
+import '../features/auth/presentation/pages/login_page.dart';
+import '../features/admin/presentation/pages/admin_dashboard.dart';
+// import '../features/doctor/presentation/pages/doctor_dashboard.dart';
+// import '../features/patient/presentation/pages/patient_dashboard.dart';
 // import fitur lain sesuai kebutuhan...
 
 final routerProvider = Provider<GoRouter>((ref) {
@@ -37,17 +37,23 @@ final routerProvider = Provider<GoRouter>((ref) {
       GoRoute(
         path: '/admin/users',
         name: 'adminUsers',
-        builder: (context, state) => Scaffold(body: Center(child: Text('Manajemen User'))),
+        builder:
+            (context, state) =>
+                Scaffold(body: Center(child: Text('Manajemen User'))),
       ),
       GoRoute(
         path: '/admin/reports',
         name: 'adminReports',
-        builder: (context, state) => Scaffold(body: Center(child: Text('Laporan Admin'))),
+        builder:
+            (context, state) =>
+                Scaffold(body: Center(child: Text('Laporan Admin'))),
       ),
       GoRoute(
         path: '/admin/settings',
         name: 'adminSettings',
-        builder: (context, state) => Scaffold(body: Center(child: Text('Pengaturan Admin'))),
+        builder:
+            (context, state) =>
+                Scaffold(body: Center(child: Text('Pengaturan Admin'))),
       ),
     ],
     // Redirect dan error handling bisa ditambahkan di sini
@@ -66,4 +72,3 @@ final routerProvider = Provider<GoRouter>((ref) {
 // Navigasi modular:
 // context.go('/adminDashboard'); // replace
 // context.push('/admin/users');  // push ke stack
-
